Rename storage layer to infrastructure
