+++ Audit-Service/src/main/java/io/github/arivanamin/lms/backend/audit/infrastructure/JpaAuditEvent.java
@@ -1,4 +1,4 @@
-package io.github.arivanamin.lms.backend.audit.storage;
+package io.github.arivanamin.lms.backend.audit.infrastructure;
 
 import io.github.arivanamin.lms.backend.core.domain.audit.AuditEvent;
 import jakarta.persistence.*;
