+++ Audit-Service/src/test/java/io/github/arivanamin/lms/backend/audit/infrastructure/JpaAuditEventTestData.java
@@ -1,4 +1,4 @@
-package io.github.arivanamin.lms.backend.audit.storage;
+package io.github.arivanamin.lms.backend.audit.infrastructure;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
