+++ Outbox-Storage/src/main/java/io/github/arivanamin/lms/backend/outbox/application/config/OutboxStorageBeansConfig.java
@@ -1,8 +1,8 @@
 package io.github.arivanamin.lms.backend.outbox.application.config;
 
 import io.github.arivanamin.lms.backend.core.domain.outbox.AuditOutboxMessageStorage;
-import io.github.arivanamin.lms.backend.outbox.storage.audit.AuditOutboxMessageRepository;
-import io.github.arivanamin.lms.backend.outbox.storage.audit.JpaAuditOutboxMessageStorage;
+import io.github.arivanamin.lms.backend.outbox.infrastructure.audit.AuditOutboxMessageRepository;
+import io.github.arivanamin.lms.backend.outbox.infrastructure.audit.JpaAuditOutboxMessageStorage;
 import org.springframework.context.annotation.*;
 
 @Configuration
