+++ Student-Service/src/main/java/io/github/arivanamin/lms/backend/student/infrastructure/DatabaseStudentStorage.java
@@ -1,12 +1,12 @@
-package io.github.arivanamin.lms.backend.student.storage;
+package io.github.arivanamin.lms.backend.student.infrastructure;
 
 import io.github.arivanamin.lms.backend.core.domain.pagination.*;
 import io.github.arivanamin.lms.backend.student.domain.entity.Student;
 import io.github.arivanamin.lms.backend.student.domain.persistence.ReadStudentsParams;
 import io.github.arivanamin.lms.backend.student.domain.persistence.StudentStorage;
-import io.github.arivanamin.lms.backend.student.storage.entity.StudentEntity;
-import io.github.arivanamin.lms.backend.student.storage.repository.StudentRepository;
-import io.github.arivanamin.lms.backend.student.storage.specification.StudentSpecification;
+import io.github.arivanamin.lms.backend.student.infrastructure.entity.StudentEntity;
+import io.github.arivanamin.lms.backend.student.infrastructure.repository.StudentRepository;
+import io.github.arivanamin.lms.backend.student.infrastructure.specification.StudentSpecification;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
