+++ Student-Service/src/main/java/io/github/arivanamin/lms/backend/student/infrastructure/entity/StudentEntity.java
@@ -1,7 +1,7 @@
-package io.github.arivanamin.lms.backend.student.storage.entity;
+package io.github.arivanamin.lms.backend.student.infrastructure.entity;
 
 import io.github.arivanamin.lms.backend.core.domain.gender.Gender;
-import io.github.arivanamin.lms.backend.outbox.storage.audit.AuditFields;
+import io.github.arivanamin.lms.backend.outbox.infrastructure.audit.AuditFields;
 import io.github.arivanamin.lms.backend.student.domain.entity.*;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.*;
