+++ Student-Service/src/test/java/io/github/arivanamin/lms/backend/student/infrastructure/JpaStudentStorageIT.java
@@ -1,9 +1,9 @@
-package io.github.arivanamin.lms.backend.student.storage;
+package io.github.arivanamin.lms.backend.student.infrastructure;
 
 import io.github.arivanamin.lms.backend.core.domain.gender.Gender;
 import io.github.arivanamin.lms.backend.student.domain.entity.Student;
-import io.github.arivanamin.lms.backend.student.storage.entity.StudentEntity;
-import io.github.arivanamin.lms.backend.student.storage.repository.StudentRepository;
+import io.github.arivanamin.lms.backend.student.infrastructure.entity.StudentEntity;
+import io.github.arivanamin.lms.backend.student.infrastructure.repository.StudentRepository;
 import io.github.arivanamin.lms.backend.testing.architecture.bases.BaseDatabaseTest;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.*;
